added admin page, term in swtds, term filter, fixed date in add swtd
