+++ src/api/admin.js
@@ -37,3 +37,44 @@ export const validateSWTD = async (data, onSuccess, onFail) => {
     onFail && onFail(error);
   }
 };
+
+export const addTerm = async (data, onSuccess, onFail) => {
+  try {
+    const response = await axios.post(
+      `http://localhost:5000/terms/`,
+      {
+        name: data.name,
+        start_date: data.start_date,
+        end_date: data.end_date,
+      },
+      {
+        headers: {
+          Authorization: `Bearer ${data.token}`,
+          "Content-Type": "application/json",
+        },
+      }
+    );
+
+    if (response.status === 200) {
+      onSuccess && onSuccess(response.data);
+    }
+  } catch (error) {
+    onFail && onFail(error);
+  }
+};
+
+export const getTerms = async (data, onSuccess, onFail) => {
+  try {
+    const response = await axios.get(`http://localhost:5000/terms/`, {
+      headers: {
+        Authorization: `Bearer ${data.token}`,
+      },
+    });
+
+    if (response.status === 200) {
+      onSuccess && onSuccess(response.data);
+    }
+  } catch (error) {
+    onFail && onFail(error);
+  }
+};
