+++ src/api/swtd.js
@@ -28,6 +28,7 @@ export const addSWTD = async (data, onSuccess, onFail, onCleanup) => {
         title: data.title,
         venue: data.venue,
         category: data.category,
+        term_id: data.term_id,
         role: data.role,
         date: data.date,
         time_started: data.time_started,
@@ -127,6 +128,7 @@ export const editSWTD = async (data, onSuccess, onFail, onCleanup) => {
         title: data.title,
         venue: data.venue,
         category: data.category,
+        term_id: data.term_id,
         role: data.role,
         date: data.date,
         time_started: data.time_started,
