+++ src/pages/dashboard/style.module.css
@@ -1,8 +1,3 @@
-.background {
-  height: 100vh;
-  overflow-y: auto;
-}
-
 .label {
   font-family: "Degradasi";
   color: #3a3238;
@@ -81,4 +76,14 @@
 .button {
   background-color: #d80085 !important;
   border: none !important;
+}
+
+.pointsBox {
+  font-family: "Poppins-SemiBold";
+  color: #d80085 !important;
+}
+
+.msg {
+  font-family: "Poppins-SemiBold";
+  color: #d80085;
 }
